refactor: useUser 훅 생성 후 리팩토링
