+++ src/pages/Detail.tsx
@@ -7,11 +7,10 @@ import IngredientBox from '../components/detail/IngredientBox';
 import StepsBox from '../components/detail/StepsBox';
 import { RecipeDataState } from '../recoil/atoms';
 import { useRecoilValue } from 'recoil';
-import { authService } from '../api/firebase';
 import useAlert from '../hooks/useAlert';
-import { User } from 'firebase/auth';
 import AlertModal from '../components/common/AlertModal';
 import CommentBox from '../components/detail/CommentBox';
+import useUser from '../hooks/useUser';
 
 const Detail = () => {
   // 레시피, 서치, 마이페이지에서 받아온 각 레시피가 가지고 있는 고유한 id
@@ -23,21 +22,8 @@ const Detail = () => {
   // 선택한 레시피를 담아줄 state
   const [recipe, setRecipe] = useState<Recipe | null>(null);
 
-  // 유저
-  const [user, setUser] = useState<User | null>(null);
-  const currentUserUid = user?.uid ?? undefined;
-
-  useEffect(() => {
-    // user 객체 존재 시 setUser 업데이트
-    const handleAuthStateChange = authService.onAuthStateChanged((user) => {
-      if (user) {
-        setUser(user);
-      }
-    });
-    return () => {
-      handleAuthStateChange();
-    };
-  }, []);
+  // 유저 상태 업데이트: useUser hook
+  const { user, currentUserUid } = useUser();
 
   // 전체 레시피와 선택한 레시피의 고유한 id가 같다면 출력
   useEffect(() => {
